Pushing updated Streams and Requirements

+++ backend/demo/demo/src/main/java/testing/demo/Controllers/UserCredentialscontroller.java
@@ -1,10 +1,8 @@
 package testing.demo.Controllers;
 
-import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import testing.demo.Services.UserCredentialsService;
-import testing.demo.model.Admin;
 import testing.demo.model.UserCredentials;
 
 import java.util.List;
@@ -36,7 +34,7 @@ public class UserCredentialscontroller {
 
 
     @DeleteMapping("/delete/{id}")
-    public String deleteById(@PathVariable int id){
+    public String deleteById(@PathVariable long id){
         return userCredentialsService.deleteById(id);
 
     }
